test: add edge case for prose with citations and fix reference-line stripping
The countWords function was stripping entire reference-style lines
(e.g. "[1]: Reber & Reeves...") because the regex treated them as
markdown link definitions. This removed real prose content from the
word count. Changed the regex to strip only the [id]: marker prefix,
preserving the citation text. Also extracted countWords into its own
module and added tests using Node's built-in test runner.
