+++ vscode-word-count/package.json
@@ -40,11 +40,13 @@
   "scripts": {
     "vscode:prepublish": "npm run compile",
     "compile": "tsc -p ./",
-    "watch": "tsc -watch -p ./"
+    "watch": "tsc -watch -p ./",
+    "test": "node --import tsx --test src/**/*.test.ts"
   },
   "devDependencies": {
-    "@types/vscode": "^1.80.0",
     "@types/node": "^20.0.0",
+    "@types/vscode": "^1.80.0",
+    "tsx": "^4.21.0",
     "typescript": "^5.4.0"
   }
 }
