+++ vscode-word-count/src/extension.ts
@@ -1,4 +1,5 @@
 import * as vscode from "vscode";
+import { countWords } from "./countWords";
 
 let statusBarItem: vscode.StatusBarItem;
 
@@ -53,28 +54,6 @@ function updateWordCount() {
 
   statusBarItem.text = label;
   statusBarItem.show();
-}
-
-function countWords(text: string): number {
-  // Strip common markdown/markup syntax so it doesn't inflate the count.
-  const cleaned = text
-    // YAML front-matter
-    .replace(/^---[\s\S]*?---/m, "")
-    // HTML tags
-    .replace(/<[^>]+>/g, "")
-    // Markdown images ![alt](url)
-    .replace(/!\[[^\]]*\]\([^)]*\)/g, "")
-    // Markdown links – keep the link text: [text](url) → text
-    .replace(/\[([^\]]*)\]\([^)]*\)/g, "$1")
-    // Markdown heading markers, blockquote markers, horizontal rules
-    .replace(/^[#>*\-=`~]+\s?/gm, "")
-    // Inline code / code fences
-    .replace(/`{1,3}[^`]*`{1,3}/g, "")
-    // Reference-style link definitions [id]: url
-    .replace(/^\[[^\]]*\]:\s.*$/gm, "");
-
-  const matches = cleaned.match(/\S+/g);
-  return matches ? matches.length : 0;
 }
 
 export function deactivate() {
